requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
tqdm>=4.66.0
PyYAML>=6.0.2
